requirements.txt
numpy
torch
Pillow
torchvision
tqdm
img2pdf
matplotlib